requirements.txt
beautifulsoup4==4.10.0
Flask==2.0.1
Flask_Cors==3.0.9
Flask_JWT_Extended==4.3.1
pandas==1.2.5
PyJWT==2.3.0
requests==2.22.0
